eidas_api/requirements.txt
PyJWT[crypto]==1.7.1
pprintpp==0.4.0
bson==0.5.10
urllib3==1.26.7
flask==2.0.1
requests==2.26.0
pycrypto==2.6.1
